pyproject.toml
[project]
name = "vessim-paper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "jupyterlab>=4.4.7",
    "nrel-pysam>=7.1.0",
    "vessim>=0.12.0",
]
